Refactor World class into partial files by functional area
Split the large World.cs file (~3100 lines) into 15 focused partial class files
organized by functionality:

- World.cs: Core fields, constructor, properties, and Dispose
- World.Entities.cs: Entity lifecycle (Spawn, Despawn, Get, Has, Add, Set, Remove)
- World.Tags.cs: String tag management
- World.Hierarchy.cs: Parent-child entity relationships
- World.Prefabs.cs: Prefab registration and instantiation
- World.Queries.cs: Query builder creation methods
- World.Systems.cs: System management (Add, Update, Enable, Disable)
- World.Singletons.cs: World-level singleton storage
- World.Validation.cs: Component validation configuration
- World.Extensions.cs: Plugin-provided API extensions
- World.Plugins.cs: Plugin lifecycle management
- World.Statistics.cs: Memory usage statistics
- World.Events.cs: Component and entity lifecycle events
- World.Messaging.cs: Immediate and queued message passing
- World.ChangeTracking.cs: Dirty flag tracking for components

This improves maintainability by grouping related functionality while preserving
all existing behavior and documentation.
