+++ src/KeenEyes.Core/World.Queries.cs
@@ -1,0 +1,50 @@
+namespace KeenEyes;
+
+public sealed partial class World
+{
+    #region Queries
+
+    /// <summary>
+    /// Creates a query for entities with the specified component.
+    /// </summary>
+    public QueryBuilder<T1> Query<T1>()
+        where T1 : struct, IComponent
+    {
+        return new QueryBuilder<T1>(this);
+    }
+
+    /// <summary>
+    /// Creates a query for entities with the specified components.
+    /// </summary>
+    public QueryBuilder<T1, T2> Query<T1, T2>()
+        where T1 : struct, IComponent
+        where T2 : struct, IComponent
+    {
+        return new QueryBuilder<T1, T2>(this);
+    }
+
+    /// <summary>
+    /// Creates a query for entities with the specified components.
+    /// </summary>
+    public QueryBuilder<T1, T2, T3> Query<T1, T2, T3>()
+        where T1 : struct, IComponent
+        where T2 : struct, IComponent
+        where T3 : struct, IComponent
+    {
+        return new QueryBuilder<T1, T2, T3>(this);
+    }
+
+    /// <summary>
+    /// Creates a query for entities with the specified components.
+    /// </summary>
+    public QueryBuilder<T1, T2, T3, T4> Query<T1, T2, T3, T4>()
+        where T1 : struct, IComponent
+        where T2 : struct, IComponent
+        where T3 : struct, IComponent
+        where T4 : struct, IComponent
+    {
+        return new QueryBuilder<T1, T2, T3, T4>(this);
+    }
+
+    #endregion
+}
