+++ src/KeenEyes.Core/World.Singletons.cs
@@ -1,0 +1,215 @@
+namespace KeenEyes;
+
+public sealed partial class World
+{
+    #region Singletons
+
+    /// <summary>
+    /// Sets or updates a singleton value for this world.
+    /// </summary>
+    /// <typeparam name="T">The singleton type. Must be a value type.</typeparam>
+    /// <param name="value">The singleton value to store.</param>
+    /// <remarks>
+    /// <para>
+    /// Singletons are world-level data not tied to any entity. They are useful for
+    /// storing global state like time, input, configuration, or other resources that
+    /// systems need to access.
+    /// </para>
+    /// <para>
+    /// If a singleton of type <typeparamref name="T"/> already exists, it will be replaced.
+    /// Use <see cref="HasSingleton{T}"/> to check existence before calling this method
+    /// if you want to avoid overwrites.
+    /// </para>
+    /// <para>
+    /// This operation is O(1) for the dictionary-based storage implementation.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// // Store game time as a singleton
+    /// world.SetSingleton(new GameTime { DeltaTime = 0.016f, TotalTime = 10.5f });
+    ///
+    /// // Update existing singleton
+    /// world.SetSingleton(new GameTime { DeltaTime = 0.016f, TotalTime = 10.516f });
+    /// </code>
+    /// </example>
+    /// <seealso cref="GetSingleton{T}"/>
+    /// <seealso cref="TryGetSingleton{T}(out T)"/>
+    /// <seealso cref="HasSingleton{T}"/>
+    /// <seealso cref="RemoveSingleton{T}"/>
+    public void SetSingleton<T>(in T value) where T : struct
+        => singletonManager.SetSingleton(in value);
+
+    /// <summary>
+    /// Gets a singleton value by reference, allowing direct modification.
+    /// </summary>
+    /// <typeparam name="T">The singleton type to retrieve.</typeparam>
+    /// <returns>A reference to the singleton data for zero-copy access.</returns>
+    /// <exception cref="InvalidOperationException">
+    /// Thrown when no singleton of type <typeparamref name="T"/> exists in this world.
+    /// Use <see cref="TryGetSingleton{T}(out T)"/> or <see cref="HasSingleton{T}"/> to check
+    /// existence before calling this method if the singleton may not exist.
+    /// </exception>
+    /// <remarks>
+    /// <para>
+    /// This method returns a reference to the boxed singleton value, enabling zero-copy
+    /// access and direct modification. Changes made through the returned reference
+    /// are immediately reflected in the stored singleton.
+    /// </para>
+    /// <para>
+    /// This operation is O(1) for the dictionary-based storage implementation.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// // Read singleton value
+    /// ref readonly var time = ref world.GetSingleton&lt;GameTime&gt;();
+    /// float delta = time.DeltaTime;
+    ///
+    /// // Modify singleton directly (zero-copy)
+    /// ref var config = ref world.GetSingleton&lt;GameConfig&gt;();
+    /// config.Difficulty = 2;
+    /// </code>
+    /// </example>
+    /// <seealso cref="SetSingleton{T}(in T)"/>
+    /// <seealso cref="TryGetSingleton{T}(out T)"/>
+    /// <seealso cref="HasSingleton{T}"/>
+    public ref T GetSingleton<T>() where T : struct
+        => ref singletonManager.GetSingleton<T>();
+
+    /// <summary>
+    /// Attempts to get a singleton value.
+    /// </summary>
+    /// <typeparam name="T">The singleton type to retrieve.</typeparam>
+    /// <param name="value">
+    /// When this method returns <c>true</c>, contains the singleton value.
+    /// When this method returns <c>false</c>, contains the default value of <typeparamref name="T"/>.
+    /// </param>
+    /// <returns>
+    /// <c>true</c> if the singleton exists; <c>false</c> otherwise.
+    /// </returns>
+    /// <remarks>
+    /// <para>
+    /// This method provides a safe way to retrieve singletons without throwing exceptions.
+    /// Unlike <see cref="GetSingleton{T}"/>, this method returns a copy of the value rather
+    /// than a reference, so modifications will not affect the stored singleton.
+    /// </para>
+    /// <para>
+    /// Use <see cref="GetSingleton{T}"/> when you need zero-copy access or want to modify
+    /// the singleton in place.
+    /// </para>
+    /// <para>
+    /// This operation is O(1) for the dictionary-based storage implementation.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// if (world.TryGetSingleton&lt;GameTime&gt;(out var time))
+    /// {
+    ///     Console.WriteLine($"Delta: {time.DeltaTime}");
+    /// }
+    /// else
+    /// {
+    ///     Console.WriteLine("GameTime singleton not set");
+    /// }
+    /// </code>
+    /// </example>
+    /// <seealso cref="GetSingleton{T}"/>
+    /// <seealso cref="HasSingleton{T}"/>
+    public bool TryGetSingleton<T>(out T value) where T : struct
+        => singletonManager.TryGetSingleton(out value);
+
+    /// <summary>
+    /// Checks if a singleton of the specified type exists in this world.
+    /// </summary>
+    /// <typeparam name="T">The singleton type to check for.</typeparam>
+    /// <returns>
+    /// <c>true</c> if a singleton of type <typeparamref name="T"/> exists;
+    /// <c>false</c> otherwise.
+    /// </returns>
+    /// <remarks>
+    /// <para>
+    /// This method provides a quick way to check singleton existence without
+    /// retrieving the value or risking exceptions.
+    /// </para>
+    /// <para>
+    /// This operation is O(1) for the dictionary-based storage implementation.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// if (world.HasSingleton&lt;GameTime&gt;())
+    /// {
+    ///     ref var time = ref world.GetSingleton&lt;GameTime&gt;();
+    ///     // Use time...
+    /// }
+    /// </code>
+    /// </example>
+    /// <seealso cref="GetSingleton{T}"/>
+    /// <seealso cref="TryGetSingleton{T}(out T)"/>
+    /// <seealso cref="SetSingleton{T}(in T)"/>
+    public bool HasSingleton<T>() where T : struct
+        => singletonManager.HasSingleton<T>();
+
+    /// <summary>
+    /// Removes a singleton from this world.
+    /// </summary>
+    /// <typeparam name="T">The singleton type to remove.</typeparam>
+    /// <returns>
+    /// <c>true</c> if the singleton was removed; <c>false</c> if no singleton
+    /// of type <typeparamref name="T"/> existed.
+    /// </returns>
+    /// <remarks>
+    /// <para>
+    /// This operation is idempotent: calling it multiple times with the same type
+    /// will return <c>false</c> after the first successful removal.
+    /// </para>
+    /// <para>
+    /// This operation is O(1) for the dictionary-based storage implementation.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// // Remove a singleton
+    /// bool removed = world.RemoveSingleton&lt;GameTime&gt;();
+    /// if (removed)
+    /// {
+    ///     Console.WriteLine("GameTime singleton removed");
+    /// }
+    ///
+    /// // Idempotent: second removal returns false
+    /// bool removedAgain = world.RemoveSingleton&lt;GameTime&gt;();
+    /// Debug.Assert(removedAgain == false);
+    /// </code>
+    /// </example>
+    /// <seealso cref="SetSingleton{T}(in T)"/>
+    /// <seealso cref="HasSingleton{T}"/>
+    public bool RemoveSingleton<T>() where T : struct
+        => singletonManager.RemoveSingleton<T>();
+
+    /// <summary>
+    /// Gets all singletons stored in this world.
+    /// </summary>
+    /// <returns>
+    /// An enumerable of tuples containing each singleton's type and boxed value.
+    /// </returns>
+    /// <remarks>
+    /// <para>
+    /// This method is primarily intended for serialization and debugging scenarios.
+    /// Values are boxed, so avoid using this in performance-critical paths.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// // Snapshot all singletons for serialization
+    /// var singletonSnapshot = world.GetAllSingletons()
+    ///     .ToDictionary(s => s.Type.FullName, s => s.Value);
+    /// </code>
+    /// </example>
+    /// <seealso cref="SetSingleton{T}(in T)"/>
+    /// <seealso cref="GetSingleton{T}"/>
+    public IEnumerable<(Type Type, object Value)> GetAllSingletons()
+        => singletonManager.GetAllSingletons();
+
+    #endregion
+}
