+++ src/KeenEyes.Core/World.Statistics.cs
@@ -1,0 +1,58 @@
+namespace KeenEyes;
+
+public sealed partial class World
+{
+    #region Memory Statistics
+
+    /// <summary>
+    /// Gets memory usage statistics for this world.
+    /// </summary>
+    /// <returns>A snapshot of current memory statistics.</returns>
+    /// <remarks>
+    /// <para>
+    /// Statistics are computed on-demand and represent a snapshot in time.
+    /// They include entity allocations, component storage, and pooling metrics.
+    /// </para>
+    /// </remarks>
+    /// <example>
+    /// <code>
+    /// var stats = world.GetMemoryStats();
+    /// Console.WriteLine($"Entities: {stats.EntitiesActive} active");
+    /// Console.WriteLine($"Archetypes: {stats.ArchetypeCount}");
+    /// Console.WriteLine($"Cache hit rate: {stats.QueryCacheHitRate:F1}%");
+    /// </code>
+    /// </example>
+    public MemoryStats GetMemoryStats()
+    {
+        // Calculate estimated component bytes
+        long estimatedBytes = 0;
+        foreach (var archetype in archetypeManager.Archetypes)
+        {
+            foreach (var componentType in archetype.ComponentTypes)
+            {
+                var info = Components.Get(componentType);
+                if (info is not null)
+                {
+                    estimatedBytes += (long)info.Size * archetype.Count;
+                }
+            }
+        }
+
+        return new MemoryStats
+        {
+            EntitiesAllocated = entityPool.TotalAllocated,
+            EntitiesActive = entityPool.ActiveCount,
+            EntitiesRecycled = entityPool.AvailableCount,
+            EntityRecycleCount = entityPool.RecycleCount,
+            ArchetypeCount = archetypeManager.ArchetypeCount,
+            ComponentTypeCount = Components.Count,
+            SystemCount = systemManager.Count,
+            CachedQueryCount = queryManager.CachedQueryCount,
+            QueryCacheHits = queryManager.CacheHits,
+            QueryCacheMisses = queryManager.CacheMisses,
+            EstimatedComponentBytes = estimatedBytes
+        };
+    }
+
+    #endregion
+}
